requirements.txt
Flask==1.0.2
Jinja2==2.10
python-dateutil==2.7.5
slackclient==1.3.0
python-dotenv==0.9.1
